pyproject.toml
[project]
name = "famigo-ui"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10, <3.11"
dependencies = [
    "alembic>=1.16.4",
    "av>=14.4.0",
    "cython>=3.1.3",
    "en-core-web-sm",
    "facenet-pytorch>=2.5.3",
    "huggingface-hub[hf-xet]>=0.34.4",
    "kokoro>=0.9.2",
    "llama-index-core>=0.13.3",
    "mediapipe>=0.10.14",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "opencv-python>=4.12.0.88",
    "pgvector>=0.4.1",
    "pillow>=11.3.0",
    "pip>=25.2",
    "playsound3>=3.2.6",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "setuptools>=80.9.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "streamlit-webrtc>=0.63.4",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "watchdog>=6.0.0",
    "wheel>=0.45.1",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
